+++ packages/cartesia/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ai-sdk/cartesia
 
+## 3.0.8
+
+### Patch Changes
+
+- Updated dependencies [0c464d9]
+- Updated dependencies [c49380c]
+  - @ai-sdk/provider-utils@5.0.14
+
+## 3.0.7
+
+### Patch Changes
+
+- Updated dependencies [1e2f324]
+  - @ai-sdk/provider@4.0.4
+  - @ai-sdk/provider-utils@5.0.13
+
 ## 3.0.6
 
 ### Patch Changes